+++ src/flepimop2/abcs/__init__.py
@@ -27,6 +27,7 @@ __all__ = [
     "EngineABC",
     "EngineProtocol",
     "ParameterABC",
+    "PatternABC",
     "ProcessABC",
     "SystemABC",
     "SystemProtocol",
@@ -35,6 +36,7 @@ __all__ = [
 from flepimop2.backend.abc import BackendABC
 from flepimop2.engine.abc import EngineABC, EngineProtocol
 from flepimop2.parameter.abc import ParameterABC
+from flepimop2.pattern.abc import PatternABC
 from flepimop2.process.abc import ProcessABC
 from flepimop2.system.abc import SystemABC
 from flepimop2.typing import SystemProtocol